qa/workunits/rgw: add tests that reproduce bucket stats inconsistency bugs
(cherry picked from commit b79dcf640ac2cc3dacf1b87bbe351db823c445d0)
